chore: code tidy
